Вынесла app в components, див для модалок перенесла в index, избавилась по коду от <></>.
