--- src/App.css
@@ -1,34 +1,0 @@
-a {
-  text-decoration: none;
-}
-
-.App {
-  width: 100%;
-  height: 100vh;
-  background-color: #131316;
-}
-
-.App__content {
-  max-width: 1240px;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  margin: 0 auto;
-  position: relative;
-}
-
-/* 88px (высота хэдера) + 40px (нижний паддинг) = 128px  */
-.App__content_errorWrapper {
-  height: calc(100% - 128px);
-}
-
-.Content__info {
-  width: 100%;
-  text-align: center;
-  position: absolute;
-  top: 50%;
-}
-
-.ModalWrapper {
-  overflow: hidden;
-}
