+++ src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
+import App from './components/app/app';
 import reportWebVitals from './reportWebVitals';
 
 import { compose, createStore, applyMiddleware } from 'redux';
@@ -23,11 +23,11 @@ const enhancer = composeEnhancers(applyMiddleware(thunk));
 const store = createStore(rootReducer, enhancer);
 
 root.render(
-  <React.StrictMode>
-    <Provider store={ store }>
-        <App />
-    </Provider>
-  </React.StrictMode>
+    <React.StrictMode>
+        <Provider store={ store }>
+            <App />
+        </Provider>
+    </React.StrictMode>
 );
 
 reportWebVitals();
